backend/requirements.txt
# Django et extensions
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0

# Base de données
mysqlclient==2.2.4
PyMySQL==1.1.0

# Machine Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# GAN et modèles avancés
tensorflow>=2.15.0
transformers>=4.35.0

# Utilitaires
python-decouple==3.8
celery==5.3.4
redis==5.0.1

# Développement
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
